+++ image_editor/core/image_processing/Image_Processing.py
@@ -1,6 +1,7 @@
 from abc import ABC
 from image_editor.core.image_processing.interface.Image_Processing_Abstract import ImageProcessingAbstract
 from image_editor.utils.convert_image import image_to_numpy_array, numpy_array_to_image, transparent_background, cv_to_image
+import image_editor as ie
 import numpy as np
 import cv2 as cv
 
@@ -12,18 +13,21 @@ class ImageProcessing(ImageProcessingAbstract, ABC):
         self.img = image_to_numpy_array(img)
         self.image_height, self.image_width, self.image_channels = self.img.shape
 
-    def extract_channel(self, channel) -> str:
+    def extract_channel(self, channel: int) -> str:
         """
         This method extracts a specific color channel
-        :param channel: 'R' / 'G' / 'B'
+        :param channel:
+            - image_editor.RED_CHANNEL
+            - image_editor.GREEN_CHANNEL
+            - image_editor.BLUE_CHANNEL
         :return: path for extracted channel of the image
         """
         new_img = self.img.copy()
-        if channel.lower() == 'r':
+        if channel == ie.RED_CHANNEL:
             new_img[:, :, [1, 2]] = 0
-        if channel.lower() == 'g':
+        if channel == ie.GREEN_CHANNEL:
             new_img[:, :, [0, 2]] = 0
-        if channel.lower() == 'b':
+        if channel == ie.BLUE_CHANNEL:
             new_img[:, :, [0, 1]] = 0
         return numpy_array_to_image(image_array=new_img, image_path=self.img_path)
 
@@ -36,13 +40,14 @@ class ImageProcessing(ImageProcessingAbstract, ABC):
         new_img = self.img.copy()
         height, width = self.img.shape[:2]
         file_extension = None
-        if kwargs.get('aspect_ratio', 'custom') == 'custom':
+        aspect_ratio = kwargs.get('aspect_ratio', ie.CROP_CUSTOM)
+        if aspect_ratio == ie.CROP_CUSTOM:
             left, top, right, bottom = kwargs.get('dimension', (0, 0, 0, 0))
             right = width - right
             bottom = height - bottom
             new_img = new_img[top:bottom, left:right, :]
-        elif kwargs.get('aspect_ratio', '1:1') == '1:1':
-            shift = kwargs.get('dimension', 0)
+        elif aspect_ratio == ie.CROP_1_1:
+            shift = kwargs.get('shift', 0)
             if width > height:
                 if (shift + height) <= height:
                     new_img = new_img[:, shift:(shift + height), :]
@@ -53,7 +58,7 @@ class ImageProcessing(ImageProcessingAbstract, ABC):
                     new_img = new_img[shift:(shift + width), :, :]
                 else:
                     new_img = new_img[(height - width):, :, :]
-        elif kwargs.get('aspect_ratio', 'circle') == 'circle':
+        elif aspect_ratio == ie.CROP_CIRCLE:
             center = kwargs.get('center', None)
             radius = kwargs.get('radius', None)
             mask = self.__create_circular_mask(height, width, center, radius)
@@ -89,5 +94,23 @@ class ImageProcessing(ImageProcessingAbstract, ABC):
 
         elif scale_amount > 0:  # downscale
             img = cv.resize(img, None, fx=scale_amount, fy=scale_amount, interpolation=cv.INTER_AREA)
+
+        return cv_to_image(image_cv=img, image_path=self.img_path)
+
+    def flip_image(self, **kwargs) -> str:
+        """
+        This method flips image horizontally / vertically based on dimensions provided
+        :param kwargs: arguments to flip the image to some dimension
+        :return: flipped image path
+        """
+        img = cv.imread(self.img_path, cv.IMREAD_COLOR)
+        flip_direction = kwargs.get('direction', ie.FLIP_HORIZONTAL)
+
+        if flip_direction == ie.FLIP_VERTICAL:
+            img = cv.flip(img, 0)
+        elif flip_direction == ie.FLIP_HORIZONTAL:
+            img = cv.flip(img, 1)
+        elif flip_direction == ie.FLIP_VERTICAL_HORIZONTAL:
+            img = cv.flip(img, -1)
 
         return cv_to_image(image_cv=img, image_path=self.img_path)