+++ image_editor/core/image_processing/interface/Image_Processing_Abstract.py
@@ -16,3 +16,7 @@ class ImageProcessingAbstract(ABC):
     @abstractmethod
     def scale_image(self, **kwargs):
         pass
+
+    @abstractmethod
+    def flip_image(self, **kwargs):
+        pass