+++ test_image.py
@@ -1,6 +1,6 @@
 from image_editor.test import test
 
 try:
-    test.test_downscale()
+    test.test_flip_vertical_horizontal()
 except Exception as e:
     print(e)